refactor: code-split VacantRoleDetailsModal behind a lazy wrapper
VacantRoleDetailsModal (~3.3k lines) was statically imported by five
components and lazily by one, which defeated the lazy import (Vite warned
the module could not be moved into its own chunk) and kept it in the main
bundle.

Add a thin VacantRoleDetailsModalLazy wrapper that dynamically imports the
modal once behind a Suspense boundary, and point the five static importers
at it (JSX and props unchanged). The modal now builds into its own chunk
(54.7 kB / 13.3 kB gzip), shrinking the main bundle by ~10.8 kB gzip, and
the Vite mixed-import warning is gone.

ESLint and production build pass; local UI smoke across the modal entry
points (cards, map, chat, invite/apply) works.
